Add files via upload
This report analyzes an observational dataset on predicting patients' creatinine clearance (CC) based on the following covariates. The data include seven variables which are: (1) "ID," which means the identity number of each patient, (2) "CC," which means creatinine clearance measurement (this is the response variable), (3) "AGE," which means the age of patients when taking the measurement, (4) "BMI" which means body mass index of each patient, (5) "SCC," which means serum creatinine concentration of each patient, (6) "SEX"  which means the sex of each patient (sex has two categories: male and female), (7) "BLOOD" which means whether a patient has high blood pressure or not. Our primary goal is to understand how these variables affect the creatinine clearance of patients.